+++ src/lib/descriptors/Material/LineBasicMaterialDescriptor.js
@@ -30,16 +30,6 @@ class LineBasicMaterialDescriptor extends MaterialDescriptorBase {
       });
     });
 
-    this.hasProp('vertexColors', {
-      type: PropTypes.oneOf([
-        THREE.NoColors,
-        THREE.FaceColors,
-        THREE.VertexColors,
-      ]),
-      simple: true,
-      default: THREE.NoColors,
-    });
-
     this.hasProp('fog', {
       type: PropTypes.bool,
       update(threeObject, fog, existsInProps) {