+++ src/lib/descriptors/Material/MeshPhongMaterialDescriptor.js
@@ -108,6 +108,17 @@ class MeshPhongMaterialDescriptor extends MaterialDescriptorBase {
         default: false,
       });
     });
+
+    this.hasMap();
+    this.hasMap('lightMap');
+    this.hasMap('aoMap');
+    this.hasMap('emissiveMap');
+    this.hasMap('bumpMap');
+    this.hasMap('normalMap');
+    this.hasMap('displacementMap');
+    this.hasMap('specularMap');
+    this.hasMap('alphaMap');
+    this.hasMap('envMap');
   }
 
   getMaterialDescription(props) {