+++ src/lib/descriptors/Material/MeshStandardMaterialDescriptor.js
@@ -103,6 +103,18 @@ class MeshStandardMaterialDescriptor extends MaterialDescriptorBase {
         default: false,
       });
     });
+
+    this.hasMap();
+    this.hasMap('lightMap');
+    this.hasMap('aoMap');
+    this.hasMap('emissiveMap');
+    this.hasMap('bumpMap');
+    this.hasMap('normalMap');
+    this.hasMap('displacementMap');
+    this.hasMap('roughnessMap');
+    this.hasMap('metalnessMap');
+    this.hasMap('alphaMap');
+    this.hasMap('envMap');
   }
 
   getMaterialDescription(props) {